[fix #1] Make list of services configurable
The OSGi module system in \BlueLaTeX 1.0 was an over-engineered
solution, and ended up being a real problem for maintainability and ease
of deployment. Still, \BlueLaTeX must remain a bit modular, its core job
being to synchronize documents in a collaborative real-time way, but the
\LaTeX compilation feature may not always be required.

So this new module system is really easy: one configures the list of
services available, each of which exposes the Rest API it handles.
There is no dynamic loading whatsoever handled, everything is loaded
when the server starts.